added authentication

+++ server/controllers/comment_controller.py
@@ -2,6 +2,7 @@ from flask import Blueprint, make_response, jsonify, request
 from flask_restful import Api, Resource, reqparse
 from server.models import Comment
 from server.config import db
+from server.auth_middleware import token_required
 
 comment_bp = Blueprint("comment_bp", __name__)
 api = Api(comment_bp)
@@ -10,7 +11,6 @@ api = Api(comment_bp)
 parser = reqparse.RequestParser()
 parser.add_argument('content', type=str, help='Provide content')
 parser.add_argument('post_id', type=int, help='Provide post id')
-parser.add_argument('user_id', type=int, help='Provide user id')
 
 
 class Comments(Resource):
@@ -18,13 +18,14 @@ class Comments(Resource):
         comment_lc = [comment.to_dict() for comment in Comment.query.all()]
         return make_response(jsonify(comment_lc), 200)
 
-    def post(self):
+    @token_required
+    def post(current_user, *args):
         try:
             args = parser.parse_args()
 
             new_comment = Comment(
                 post_id=args["post_id"],
-                user_id=args["user_id"],
+                user_id=current_user.id,
                 content=args["content"]
             )
             db.session.add(new_comment)
@@ -43,11 +44,24 @@ class CommentByID(Resource):
         response = make_response(jsonify(comment.to_dict()), 200)
         return response
 
-    def patch(self, comment_id):
+    @token_required
+    def patch(current_user, *args, comment_id):
+        comment = Comment.query.filter_by(id=comment_id).first()
+        if not comment or comment.user_id != current_user.id:
+            return {
+                "message": "failed to update comment",
+                "error": "Unauthorized request",
+                "data": None
+            }, 401
         try:
-            comment = Comment.query.filter_by(id=comment_id).first()
             data = request.get_json()
             for attr in data:
+                if attr in ['post_id', 'id']:
+                    return {
+                        "message": "failed to update comment. You can only edit the content",
+                        "error": "Unauthorized request",
+                        "data": None
+                    }, 401
                 setattr(comment, attr, data[attr])
 
             db.session.commit()
@@ -57,8 +71,15 @@ class CommentByID(Resource):
         except ValueError as e:
             return {"error": [str(e)]}
 
-    def delete(self, comment_id):
+    @token_required
+    def delete(current_user, *args, comment_id):
         comment = Comment.query.filter_by(id=comment_id).first()
+        if not comment or comment.user_id != current_user.id:
+            return {
+                "message": "failed to delete comment",
+                "error": "Unauthorized request",
+                "data": None
+            }, 401
 
         db.session.delete(comment)
         db.session.commit()
